基本的视图使用
