--- Project/djangoProject/post/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-from django.contrib import admin
-
-from .models import ImagesPost
-
-admin.site.register(ImagesPost)
