--- Project/djangoProject/requirements.txt
@@ -1,2 +1,0 @@
-django==3.2.14
-djangorestframework==3.13.1
